Controls - Added some more potential cases for the key listener, potential way for ctrl-keycode to work.
GUI - implemented the menu items and the action listeners so that when they are selected they will be able to invoke a method that will be of use. Link to other players components (save, load, etc.).

HelpMenuListener - class created, needed as JMenu do not have an action listener, created my own which will prompt the dialog box with all the helpful instructions for the game.